requirements.txt
discord.py==2.3.2
aiosqlite
aiohttp
pip install requests
